+++ src/App.tsx
@@ -6,11 +6,14 @@ import { UserWarning } from './UserWarning';
 import { getTodos, USER_ID } from './api/todos';
 import { Todo } from './types/Todo';
 import classNames from 'classnames';
+import { Header } from './components/Header';
+import { TodoList } from './components/TodoList';
+import { Footer } from './components/Footer';
 // #endregion
 
 export const App: React.FC = () => {
   // #region useState
-  const [inputValue, setInputValue] = useState('');
+
   const [todos, setTodos] = useState<Todo[]>([]);
   const [allTodos, setAllTodos] = useState<Todo[]>([]);
   const [selectedTodos, setSelectedTodos] = useState<number[]>([]);
@@ -30,9 +33,11 @@ export const App: React.FC = () => {
   }, []);
 
   useEffect(() => {
-    const notCompleted = allTodos.filter(todo => !todo.completed);
+    if (allTodos) {
+      const notCompleted = allTodos.filter(todo => !todo.completed);
 
-    setTodosCounter(notCompleted.length);
+      setTodosCounter(notCompleted.length);
+    }
   }, [selectedTodos, todos, allTodos]);
 
   useEffect(() => {
@@ -65,150 +70,25 @@ export const App: React.FC = () => {
       <h1 className="todoapp__title">todos</h1>
 
       <div className="todoapp__content">
-        <header className="todoapp__header">
-          {/* this button should have `active` class only if all todos are completed */}
-          <button
-            type="button"
-            className="todoapp__toggle-all active"
-            data-cy="ToggleAllButton"
-          />
+        <Header />
 
-          {/* Add a todo on form submit */}
-          <form
-            onSubmit={event => {
-              event.preventDefault();
-              setInputValue('');
-            }}
-          >
-            <input
-              data-cy="NewTodoField"
-              type="text"
-              className="todoapp__new-todo"
-              placeholder="What needs to be done?"
-              value={inputValue}
-              onChange={e => setInputValue(e.target.value)}
-            />
-          </form>
-        </header>
-
-        <section className="todoapp__main" data-cy="TodoList">
-          {/* This is a completed todo */}
-          {todos.map(todo => {
-            return (
-              <div
-                data-cy="Todo"
-                className={classNames('todo', {
-                  completed: selectedTodos.includes(todo.id) || todo.completed,
-                })}
-                key={todo.id}
-              >
-                <label className="todo__status-label">
-                  <input
-                    data-cy="TodoStatus"
-                    type="checkbox"
-                    className={classNames('todo__status')}
-                    checked={todo.completed}
-                    onChange={() => {
-                      //toggle completed or not todo
-                      if (selectedTodos.includes(todo.id) || todo.completed) {
-                        const filteredTodos = selectedTodos.filter(
-                          item => item !== todo.id,
-                        );
-
-                        setSelectedTodos(filteredTodos);
-                      } else {
-                        setSelectedTodos(prev => [...prev, todo.id]);
-                      }
-                    }}
-                  />
-                </label>
-
-                <span data-cy="TodoTitle" className="todo__title">
-                  {todo.title}
-                </span>
-                <button
-                  type="button"
-                  className="todo__remove"
-                  data-cy="TodoDelete"
-                  onClick={() => {
-                    const deletedTodo = todo.id;
-                    const filteredList = todos.filter(
-                      todoItem => todoItem.id !== deletedTodo,
-                    );
-
-                    setAllTodos(filteredList);
-                  }}
-                >
-                  ×
-                </button>
-
-                <div data-cy="TodoLoader" className="modal overlay">
-                  <div className="modal-background has-background-white-ter" />
-                  <div className="loader" />
-                </div>
-              </div>
-            );
-          })}
-         
-        </section>
+        {/* This is a completed todo */}
+        <TodoList
+          todos={todos}
+          selectedTodos={selectedTodos}
+          setSelectedTodos={setSelectedTodos}
+          setAllTodos={setAllTodos}
+        />
 
         {/* Hide the footer if there are no todos */}
         {allTodos.length > 0 && (
-          <footer className="todoapp__footer" data-cy="Footer">
-            <span className="todo-count" data-cy="TodosCounter">
-              {`${todosCounter} items left`}
-            </span>
-
-            {/* Active link should have the 'selected' class */}
-            <nav className="filter" data-cy="Filter">
-              <a
-                href="#/"
-                className={classNames('filter__link', {
-                  selected: selectedLink === 'All',
-                })}
-                data-cy="FilterLinkAll"
-                onClick={() => setSelectedLink('All')}
-              >
-                All
-              </a>
-
-              <a
-                href="#/active"
-                className={classNames('filter__link', {
-                  selected: selectedLink === 'active',
-                })}
-                data-cy="FilterLinkActive"
-                onClick={() => setSelectedLink('active')}
-              >
-                Active
-              </a>
-
-              <a
-                href="#/completed"
-                className={classNames('filter__link', {
-                  selected: selectedLink === 'completed',
-                })}
-                data-cy="FilterLinkCompleted"
-                onClick={() => setSelectedLink('completed')}
-              >
-                Completed
-              </a>
-            </nav>
-
-            {/* this button should be disabled if there are no completed todos */}
-            <button
-              type="button"
-              className="todoapp__clear-completed"
-              data-cy="ClearCompletedButton"
-              onClick={() => {
-                const activeTodos = todos.filter(todo => !todo.completed);
-
-                setAllTodos(activeTodos);
-              }}
-            >
-              Clear completed
-            </button>
-          </footer>
+          <Footer
+            todosCounter={todosCounter}
+            selectedLink={selectedLink}
+            setSelectedLink={setSelectedLink}
+            todos={todos}
+            setAllTodos={setAllTodos}
+          />
         )}
       </div>
