+++ src/components/Header.tsx
@@ -1,0 +1,33 @@
+import { useState } from 'react';
+
+export const Header = () => {
+  const [inputValue, setInputValue] = useState('');
+
+  return (
+    <header className="todoapp__header">
+      {/* this button should have `active` class only if all todos are completed */}
+      <button
+        type="button"
+        className="todoapp__toggle-all active"
+        data-cy="ToggleAllButton"
+      />
+
+      {/* Add a todo on form submit */}
+      <form
+        onSubmit={event => {
+          event.preventDefault();
+          setInputValue('');
+        }}
+      >
+        <input
+          data-cy="NewTodoField"
+          type="text"
+          className="todoapp__new-todo"
+          placeholder="What needs to be done?"
+          value={inputValue}
+          onChange={e => setInputValue(e.target.value)}
+        />
+      </form>
+    </header>
+  );
+};
